Tier B Phase 2 (Slice 3): host-app-bridge → @mcpjam/sdk/widget-runtime
Move the framework-free MCP Apps host bridge surface (createHostAppBridge,
registerHostBridgeHandlers, and the app-tool invocation lifecycle types) into
the @mcpjam/sdk/widget-runtime subpath so the production renderer and the eval
browser harness share one source of truth.

This slice adds a real @modelcontextprotocol/ext-apps dependency because the
host bridge needs AppBridge as a runtime value. To keep the SDK on NodeNext and
the published .d.ts clean for external NodeNext consumers, the module imports
only the AppBridge value and derives the ext-apps host types from the AppBridge
constructor signature — importing those types by name is impossible under
NodeNext because ext-apps/app-bridge re-exports them through an extensionless
`export * from "./types"` the resolver cannot follow. Mirrors the existing
derivation of the bridge-handler param/result types.

The inspector's host-app-bridge.ts and app-tool-invocations.ts become thin
back-compat re-export shims, so existing import paths (and the convenience
re-exports of the tool-visibility + iframe-sandbox helpers) are unchanged.
No behavior change.
