+++ mcpjam-inspector/client/src/components/chat-v2/thread/app-tool-invocations.ts
@@ -1,15 +1,16 @@
-export type AppToolInvocationStatus = "running" | "success" | "error";
+/**
+ * app-tool-invocations.ts — back-compat re-export shim.
+ *
+ * The app-tool invocation lifecycle types now live in
+ * `@mcpjam/sdk/widget-runtime` (Tier B Phase 2) so the host bridge that emits
+ * them and the inspector renderer UI that displays them share one definition.
+ * This file preserves the existing
+ * `@/components/chat-v2/thread/app-tool-invocations` import path used across the
+ * thread renderer (transcript, replay, parts, message-view, part-switch).
+ */
 
-export interface AppToolInvocation {
-  id: string;
-  parentToolCallId: string;
-  toolName: string;
-  input?: Record<string, unknown>;
-  output?: unknown;
-  errorText?: string;
-  status: AppToolInvocationStatus;
-  startedAt: number;
-  completedAt?: number;
-}
-
-export type AppToolInvocationUpdate = AppToolInvocation;
+export type {
+  AppToolInvocation,
+  AppToolInvocationStatus,
+  AppToolInvocationUpdate,
+} from "@mcpjam/sdk/widget-runtime";
